fix breakage from PR #127
